requirements.txt
pydantic
requests
python-dotenv
kafka-python